+++ src/contacts/mod.rs
@@ -5,9 +5,12 @@ mod contact;
 mod contact_email;
 mod contacts_data;
 mod contacts_metadata;
+mod contacts_tool;
 mod phone_number;
+mod queries;
 mod social_profile;
 
+pub use contacts_tool::ContactsTool;
 pub use address::Address;
 pub use contact::Contact;
 pub use contact_email::ContactEmail;